Use torch_dtype kwarg for AutoModel.from_pretrained
The 'dtype=' alias only exists in transformers >=4.56; the prod image
ships an older version that raises TypeError. 'torch_dtype' has been
the canonical kwarg for years and is backward compatible.

+++ backend/CohortVarLinker/updated_CohortVarLinker/embed_model.py
@@ -122,7 +122,7 @@ class UnifiedEmbeddingModel:
             model_name_or_path, cache_dir=cache, trust_remote_code=True)
         self.model = AutoModel.from_pretrained(
             model_name_or_path, cache_dir=cache, trust_remote_code=True,
-            dtype=dtype)
+            torch_dtype=dtype)
 
         if self._is_decoder:
             self.tokenizer.padding_side = "left"
